break out specifics from comon
