+++ local-cli/rnpm/windows/src/wpf.js
@@ -11,6 +11,18 @@
 
 const Common = require('./common');
 const chalk = require('chalk');
+const execSync = require('child_process').execSync;
+const path = require('path');
+
+const REACT_NATIVE_WPF_GENERATE_PATH = function() {
+  return path.resolve(
+    process.cwd(),
+    'node_modules',
+    'react-native-windows',
+    'local-cli',
+    'generate-wpf.js'
+  );
+};
 
 module.exports = function windows(config, args, options) {
   const name = args[0] ? args[0] : Common.getReactNativeAppName();
@@ -20,10 +32,10 @@ module.exports = function windows(config, args, options) {
   return Common.getInstallPackage(version)
     .then(rnwPackage => {
       console.log(`Installing ${rnwPackage}...`);
-      // execSync(`npm install --save ${rnwPackage}`);
+      execSync(`npm install --save ${rnwPackage}`);
       console.log(chalk.green(`${rnwPackage} successfully installed.`));
 
-      const generateWindows = require(Common.REACT_NATIVE_WPF_GENERATE_PATH());
+      const generateWindows = require(REACT_NATIVE_WPF_GENERATE_PATH());
       generateWindows(process.cwd(), name, ns);
     }).catch(error => console.error(chalk.red(error.message)));
 }
